requirements.txt
python-telegram-bot==20.2
openai==0.27.4
langchain==0.0.147
pydub==0.25.1
python-dotenv==1.0.0
loguru==0.7.0
click==8.1.3